requirements.txt
streamlit==1.38.0
huggingface_hub==0.25.1
python-dotenv==1.0.1